etl-nutrition/requirements.txt
pandas>=2.2.0
requests>=2.31.0
pydantic>=2.7.0
SQLAlchemy>=2.0.0
psycopg[binary,pool]>=3.2.0
prefect>=2.14.0
pyyaml>=6.0.0
